+++ src/components/RegionSelector/RegionSelector.tsx
@@ -1,6 +1,6 @@
 /// <reference types="googlemaps" />
 import { GoogleMap, Polygon, useLoadScript } from "@react-google-maps/api";
-import React, { useMemo, useState } from "react";
+import React, { useCallback, useMemo, useState } from "react";
 import { Message, Segment } from "semantic-ui-react";
 import convert from "./conversion";
 import styles from "./RegionSelector.module.css";
@@ -15,12 +15,6 @@ const options: google.maps.MapOptions = {
 
 interface RegionSelectorProps {
   value: string;
-  center: {
-    lat: number;
-    lng: number;
-  };
-  latLngBounds?: google.maps.LatLngBounds;
-  zoom: number;
   onChange: (event: string) => void;
 }
 
@@ -35,18 +29,26 @@ export default function RegionSelector(props: RegionSelectorProps) {
     googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_KEY!,
   });
 
-  function onPolygonChange() {
+  const onPolygonChange = useCallback(() => {
     props.onChange(convert.polygonToString(polygonRef!));
-  }
+  }, [polygonRef, props.onChange]);
 
-  function onPolygonRemove(event: google.maps.PolyMouseEvent) {
-    if (event.vertex === undefined) {
-      return;
-    }
+  const onLoad = useCallback(
+    (map) => polygonPath && map.fitBounds(toLatLngBounds(polygonPath)),
+    [polygonPath],
+  );
 
-    polygonRef!.getPath().removeAt(event.vertex);
-    onPolygonChange();
-  }
+  const onPolygonRemove = useCallback(
+    (event: google.maps.PolyMouseEvent) => {
+      if (event.vertex === undefined) {
+        return;
+      }
+
+      polygonRef!.getPath().removeAt(event.vertex);
+      onPolygonChange();
+    },
+    [polygonRef, onPolygonChange],
+  );
 
   if (!isLoaded) {
     return <Segment loading className={styles.map} />;
@@ -68,14 +70,7 @@ export default function RegionSelector(props: RegionSelectorProps) {
     <GoogleMap
       options={options}
       mapContainerClassName={styles.map}
-      center={props.center}
-      zoom={13}
-      onLoad={(map) => {
-        // TODO: this could be derived from polygonPath
-        if (props.latLngBounds) {
-          map.fitBounds(props.latLngBounds);
-        }
-      }}
+      onLoad={onLoad}
     >
       <Polygon
         draggable
@@ -88,4 +83,12 @@ export default function RegionSelector(props: RegionSelectorProps) {
       />
     </GoogleMap>
   );
+}
+
+function toLatLngBounds(
+  points: { lng: number; lat: number }[],
+): google.maps.LatLngBounds {
+  const bounds = new google.maps.LatLngBounds();
+  points.forEach((point) => bounds.extend(point));
+  return bounds;
 }
